+++ src/com/sun/jna/NativeLibrary.java
@@ -257,11 +257,14 @@ public class NativeLibrary {
             if (handle == 0) {
                 try {
                     File embedded = Native.extractFromResourcePath(libraryName, (ClassLoader)options.get(Library.OPTION_CLASSLOADER));
-                    handle = Native.open(embedded.getAbsolutePath(), openFlags);
-                    libraryPath = embedded.getAbsolutePath();
-                    // Don't leave temporary files around
-                    if (Native.isUnpacked(embedded)) {
-                        Native.deleteLibrary(embedded);
+                    try {
+                        handle = Native.open(embedded.getAbsolutePath(), openFlags);
+                        libraryPath = embedded.getAbsolutePath();
+                    } finally {
+                        // Don't leave temporary files around
+                        if (Native.isUnpacked(embedded)) {
+                            Native.deleteLibrary(embedded);
+                        }
                     }
                 }
                 catch(IOException e2) { e = new UnsatisfiedLinkError(e2.getMessage()); }